+++ CHANGELOG.md
@@ -45,6 +45,7 @@ Yii Core version 2 Change Log
 - Bug #104: Remove `insert()` method from `Schema::class` (@terabytesoftw)
 - Bug #105: Removed the `sequenceName` property from the `TableSchema::class` and added it to the `ColumnSchema::class` (@terabytesoftw)
 - Enh #106: Move `extractAlias()` method to `Quoter::class` (@terabytesoftw)
+- Enh #107: Add `SqlHelper::class` in `db` (@terabytesoftw)
 
 Yii Framework 2 Change Log
 ==========================